+++ kadai1/shimastripe/imgconv/imgconv.go
@@ -1,4 +1,4 @@
-package shimastripe
+package imgconv
 
 import (
 	"errors"
